nextauth-app/ocr-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
google-cloud-vision==3.7.2
